# perf(ext/fetch): use new instead of createBranded