--- a/ext/fetch/20_headers.js
+++ b/ext/fetch/20_headers.js
@@ -40,6 +40,7 @@ const _headerList = Symbol("header list");
 const _iterableHeaders = Symbol("iterable headers");
 const _iterableHeadersCache = Symbol("iterable headers cache");
 const _guard = Symbol("guard");
+const _brand = webidl.brand;
 
 /**
  * @typedef Header
@@ -286,8 +287,8 @@ class Headers {
 
   /** @param {HeadersInit} [init] */
   constructor(init = undefined) {
-    if (init === webidl.brand) {
-      this[init] = init;
+    if (init === _brand) {
+      this[_brand] = _brand;
       return;
     }
 
@@ -296,7 +297,7 @@ class Headers {
       init = webidl.converters["HeadersInit"](init, prefix, "Argument 1");
     }
 
-    this[webidl.brand] = webidl.brand;
+    this[_brand] = _brand;
     this[_guard] = "none";
     if (init !== undefined) {
       fillHeaders(this, init);
@@ -491,7 +492,7 @@ webidl.converters["Headers"] = webidl.createInterfaceConverter(
  * @returns {Headers}
  */
 function headersFromHeaderList(list, guard) {
-  const headers = new Headers(webidl.brand);
+  const headers = new Headers(_brand);
   headers[_headerList] = list;
   headers[_guard] = guard;
   return headers;