--- a/ext/fetch/23_response.js
+++ b/ext/fetch/23_response.js
@@ -306,7 +306,7 @@ class Response {
    */
   constructor(body = null, init = undefined) {
     if (body === webidl.brand) {
-      this[webidl.brand] = webidl.brand;
+      this[body] = body;
       return;
     }